Move use cases into directories for consistency
Jira ticket: CAMS-510
