+++ backend/lib/controllers/attorneys/attorneys.controller.ts
@@ -1,5 +1,5 @@
 import { ApplicationContext } from '../../adapters/types/basic';
-import AttorneysList from '../../use-cases/attorneys';
+import AttorneysList from '../../use-cases/attorneys/attorneys';
 import { AttorneyUser } from '../../../../common/src/cams/users';
 import { CamsHttpResponseInit, httpSuccess } from '../../adapters/utils/http-response';
 import { getCamsError } from '../../common-errors/error-utilities';
