+++ backend/lib/controllers/case-summary/case-summary.controller.test.ts
@@ -1,5 +1,5 @@
 import { createMockApplicationContext } from '../../testing/testing-utilities';
-import CaseManagement from '../../use-cases/case-management';
+import CaseManagement from '../../use-cases/cases/case-management';
 import { NotFoundError } from '../../common-errors/not-found-error';
 import { MockData } from '../../../../common/src/cams/test-utilities/mock-data';
 import { CaseSummaryController } from './case-summary.controller';
