+++ backend/lib/controllers/case-summary/case-summary.controller.ts
@@ -1,6 +1,6 @@
 import { ApplicationContext } from '../../adapters/types/basic';
 import { CaseSummary } from '../../../../common/src/cams/cases';
-import CaseManagement from '../../use-cases/case-management';
+import CaseManagement from '../../use-cases/cases/case-management';
 import { CamsHttpResponseInit, httpSuccess } from '../../adapters/utils/http-response';
 import { getCamsError } from '../../common-errors/error-utilities';
 import { CamsController } from '../controller';
