+++ backend/lib/controllers/cases/cases.controller.test.ts
@@ -1,7 +1,7 @@
 import { MockData } from '../../../../common/src/cams/test-utilities/mock-data';
 import { ApplicationContext } from '../../adapters/types/basic';
 import { CamsHttpResponseInit } from '../../adapters/utils/http-response';
-import CaseManagement from '../../use-cases/case-management';
+import CaseManagement from '../../use-cases/cases/case-management';
 import { CasesController } from './cases.controller';
 import { CaseBasics, CaseDetail } from '../../../../common/src/cams/cases';
 import {
