+++ backend/lib/use-cases/attorneys/attorney.gateway.interface.d.ts
@@ -1,5 +1,5 @@
-import { ApplicationContext } from '../adapters/types/basic';
-import { AttorneyUser } from '../../../common/src/cams/users';
+import { ApplicationContext } from '../../adapters/types/basic';
+import { AttorneyUser } from '../../../../common/src/cams/users';
 
 export interface AttorneyGatewayInterface {
   getAttorneys(applicationContext: ApplicationContext): Promise<Array<AttorneyUser>>;
