+++ backend/lib/use-cases/case-assignment/assignment.exception.ts
@@ -1,5 +1,5 @@
-import { CamsError, CamsErrorOptions } from '../common-errors/cams-error';
-import HttpStatusCodes from '../../../common/src/api/http-status-codes';
+import { CamsError, CamsErrorOptions } from '../../common-errors/cams-error';
+import HttpStatusCodes from '../../../../common/src/api/http-status-codes';
 
 /* eslint-disable-next-line @typescript-eslint/no-empty-object-type */
 export interface AssignmentErrorOptions extends CamsErrorOptions {}
