+++ backend/lib/use-cases/orders/orders-consolidation-approval.test.ts
@@ -12,7 +12,7 @@ import { MockData } from '../../../../common/src/cams/test-utilities/mock-data';
 import { ApplicationContext } from '../../adapters/types/basic';
 import * as crypto from 'crypto';
 import { CaseHistory, ConsolidationOrderSummary } from '../../../../common/src/cams/history';
-import { CaseAssignmentUseCase } from '../case-assignment';
+import { CaseAssignmentUseCase } from '../case-assignment/case-assignment';
 import { CamsRole } from '../../../../common/src/cams/roles';
 import { SYSTEM_USER_REFERENCE } from '../../../../common/src/cams/auditable';
 import { REGION_02_GROUP_NY } from '../../../../common/src/cams/test-utilities/mock-user';
